changes to support airplane profiles

+++ DragAndDropExtractor/ArchiveExtractor.cs
@@ -8,7 +8,7 @@ using SharpCompress.Archives;
 
 namespace DragAndDropInstaller;
 
-internal class ArchiveExtractor(string destinationPath)
+internal class ArchiveExtractor()
 {
     private readonly List<IArchiveEntry> toExtract = [];
     private readonly List<IArchiveEntry> DotPyFiles = [];
@@ -16,6 +16,8 @@ internal class ArchiveExtractor(string destinationPath)
     private readonly List<IArchiveEntry> DotCfgFiles = [];
     private readonly List<string> deletedFiles = [];
     private readonly List<string> installedFiles = [];
+    private readonly string profilesPath = Path.Combine(Path.GetFullPath(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)), "Virtuali", "GSX", "MSFS");
+    private readonly string airplanesPath = Path.Combine(Path.GetFullPath(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)), "Virtuali", "Airplanes");
     private bool multipleProfileFound;
 
     public void ExtractFiles(string archivePath)
@@ -45,31 +47,69 @@ internal class ArchiveExtractor(string destinationPath)
 
         HandleMultipleProfiles(DotIniFiles);
         HandleMultipleProfiles(DotPyFiles);
+        HandleMultipleProfiles(DotCfgFiles);
         Stopwatch sw = Stopwatch.StartNew();
-        foreach (IArchiveEntry entry in toExtract)
-        {
-            //needs to be modified for cfg files (cfg files needs to be extract with its folder)
-            string fullDestinationPath = Path.GetFullPath(Path.Combine(destinationPath, GetFileName(entry.Key)));
-            installedFiles.Add(fullDestinationPath);
-            using Stream stream = entry.OpenEntryStream();
-            using FileStream writer = File.OpenWrite(fullDestinationPath);
-            stream.CopyTo(writer);
-        }
+        Extract();
         sw.Stop();
         UserInterface.DisplayChanges(installedFiles, deletedFiles);
         UserInterface.DisplayElapsedTime(sw);
     }
+    private void Extract()
+    {
+        //needs to be modified for cfg files (cfg files needs to be extract with its folder)
+        foreach (IArchiveEntry entry in toExtract)
+        {
+            if (entry.Key.EndsWith(".cfg", StringComparison.OrdinalIgnoreCase))
+            {
+                string ContainingDirectory;
+                if (entry.Key.Contains('\\'))
+                {
+                    
+                    ContainingDirectory = entry.Key.Split('\\')[entry.Key.Split('\\').Length - 2];
+                }
+                else
+                {
+                     ContainingDirectory = entry.Key.Split('/')[entry.Key.Split('/').Length - 2];
+                }
+                string PathToDirectory = Path.GetFullPath(Path.Combine(airplanesPath, ContainingDirectory));
+                if (!Directory.Exists(PathToDirectory))
+                {
+                    Directory.CreateDirectory(PathToDirectory);
+                }
+                string fullDestinationPath = Path.GetFullPath(Path.Combine(airplanesPath, ContainingDirectory, GetFileName(entry.Key)));
+                installedFiles.Add(fullDestinationPath);
+                using Stream stream = entry.OpenEntryStream();
+                using FileStream writer = File.OpenWrite(fullDestinationPath);
+                stream.CopyTo(writer);
+            }
+            else
+            {
+                string fullDestinationPath = Path.GetFullPath(Path.Combine(profilesPath, GetFileName(entry.Key)));
+                installedFiles.Add(fullDestinationPath);
+                using Stream stream = entry.OpenEntryStream();
+                using FileStream writer = File.OpenWrite(fullDestinationPath);
+                stream.CopyTo(writer);
+            }
+
+        }
+    }
     private void HandleSupportedFile(List<IArchiveEntry> list, IArchiveEntry entry)
     {
-        //needs to be modified for cfg files, there should be no deletion if its a .cfg file
+        //needs to be modified for cfg files, there should be no deletion if its a .cfg file. But somehow we should indicate, if the profile is overwritten?
+        if (entry.Key.EndsWith(".cfg", StringComparison.OrdinalIgnoreCase))
+        {
+            list.Add(entry);
+            return;
+        }    
         list.Add(entry);
-        string[] matchingFiles = Directory.GetFiles(destinationPath, $"*{GetICAOcode(entry.Key)}*");
+        string[] matchingFiles = Directory.GetFiles(profilesPath, $"*{GetICAOcode(entry.Key)}*");
         deletedFiles.AddRange(matchingFiles);
         matchingFiles.ForEach(filePath => File.Delete(filePath));
     }
 
     private void HandleMultipleProfiles(List<IArchiveEntry> list)
     {
+        //Multiple selection for .CFG files
         if (list.Count <= 1)
         {
             if (list.Count == 1)
